+++ src/main/java/com/watsonsoftware/document/repository/DocumentRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 public interface DocumentRepository extends
         JpaRepository<DocumentEntity, Integer> {
 
-    DocumentEntity findOneByDocIdAndOwnerId(Long docId, String ownerId);
+    DocumentEntity findOneByDocIdAndOwnerId(Integer docId, String ownerId);
 
 }
